changing int parameters to float and fixing mistake in Grid class

+++ input_output.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 using namespace libconfig;
 
-int load_config(int *a, int *b, int *i_max, int *j_max, int *boundary_condition, float *u_in, float *v_in, int *Re, float *tau, float *g_x, float *g_y)
+int load_config(float *a, float *b, int *i_max, int *j_max, int *boundary_condition, float *u_in, float *v_in, float *Re, float *tau, float *g_x, float *g_y)
 {   
     Config cfg;
