+++ input_output.h
@@ -1,7 +1,7 @@
 #ifndef INPUT_OUTPUT_H
 #define INPUT_OUTPUT_H
 
-int load_config(int *a, int *b, int *i_max, int *j_max, int *boundary_condition, float *u_in, float *v_in, int *Re, float *tau, float *g_x, float *g_y);
+int load_config(float *a, float *b, int *i_max, int *j_max, int *boundary_condition, float *u_in, float *v_in, float *Re, float *tau, float *g_x, float *g_y);
 
 void print_grid(float ** u);
